requirements.txt
torch
transformers
sentence-transformers
fastapi
uvicorn
gunicorn
PyJWT